Extract shared FindLocalPProject helper for CLI options
PCheckerOptions and PCompilerOptions both contained near-identical
implementations of FindLocalPProject that scanned the current
directory for a *.pproj file. The two had silently drifted:

- PCompilerOptions guarded against double-resolution when the user
  already passed --pproj or --pfiles; PCheckerOptions did not, so
  running 'p check --pproj foo.pproj' would still trigger a directory
  scan and could append a second pproj argument from CWD.
- PCompilerOptions emitted the search/no-match/found log lines via
  CommandLineOutput; the same calls in PCheckerOptions had been
  commented out, suppressing diagnostics on the check path.

Move the canonical implementation (with the guard and the logging) to
a new ProjectFileLocator helper in Plang.Options, and have both
option parsers call it. No behavior change for 'p compile'; 'p check'
picks up the early-exit guard and the previously-suppressed log
lines.
