+++ Src/PCompiler/PCommandLine/Options/PCompilerOptions.cs
@@ -65,7 +65,7 @@ namespace Plang.Options
             {
                 var result = Parser.ParseArguments(args);
                 // if there are no --pproj or --pfiles arguments, then search for a pproj file locally and load it
-                FindLocalPProject(result);
+                ProjectFileLocator.FindLocalPProject(result);
 
                 // load pproj file first
                 UpdateConfigurationWithPProjectFile(compilerConfiguration, result);
@@ -98,37 +98,6 @@ namespace Plang.Options
             }
 
             return compilerConfiguration;
-        }
-
-        private static void FindLocalPProject(List<CommandLineArgument> result)
-        {
-            foreach (var arg in result)
-            {
-                if (arg.LongName.Equals("pproj") || arg.LongName.Equals("pfiles"))
-                    return;
-            }
-
-            CommandLineOutput.WriteInfo(".. Searching for a P project file *.pproj locally in the current folder");
-            var filtered =
-                from file in Directory.GetFiles(Directory.GetCurrentDirectory(), "*.pproj")
-                let info = new FileInfo(file)
-                where (((info.Attributes & FileAttributes.Hidden) ==0)& ((info.Attributes & FileAttributes.System)==0))
-                select file;
-            var files = filtered.ToArray();
-            if (files.Length == 0)
-            {
-                CommandLineOutput.WriteInfo(
-                    $".. No P project file *.pproj found in the current folder: {Directory.GetCurrentDirectory()}");
-            }
-            else
-            {
-                var commandlineArg = new CommandLineArgument();
-                commandlineArg.Value = files.First();
-                commandlineArg.LongName = "pproj";
-                commandlineArg.ShortName = "pp";
-                CommandLineOutput.WriteInfo($".. Found P project file: {commandlineArg.Value}");
-                result.Add(commandlineArg);
-            }
         }
 
         /// <summary>
